Fixed numerous bugs. Separated various classes and data into separate files Moved the python code into its own directory
