+++ example_code/README
@@ -1,0 +1,1 @@
+headers_sample.py is what you want to look at. The rest exists to support that.
