+++ example_code/default_headers.py
@@ -1,0 +1,47 @@
+
+default_requests = [
+  {':method': "get",
+    ':path': '/index.html',
+    ':version': 'HTTP/1.1',
+    'user-agent': 'blah blah browser version blah blah',
+    'accept-encoding': 'sdch, bzip, compress',
+    ':host': 'www.foo.com',
+    'referrer': 'www.foo.com/ILoveBugs',
+    'cookie': 'SOMELONGSTRINGTHATISMOSTLYOPAQUE;BLAJBLA',
+    'date': 'Wed Jul 18 11:50:43 2012'},
+  {':method': "get",
+    ':path': '/index.js',
+    ':version': 'HTTP/1.1',
+    'user-agent': 'blah blah browser version blah blah',
+    'accept-encoding': 'sdch, bzip, compress',
+    ':host': 'www.foo.com',
+    'referrer': 'www.foo.com/ILoveBugsALot',
+    'cookie': 'SOMELONGSTRINGTHATISMOSTLYOPAQUE;BLAJBLA',
+    'date': 'Wed Jul 18 11:50:44 2012'},
+  {':method': "get",
+    ':path': '/index.css',
+    ':version': 'HTTP/1.1',
+    'user-agent': 'blah blah',
+    'accept-encoding': 'sdch, bzip, compress',
+    ':host': 'www.foo.com',
+    'cookie': 'SOMELONGSTRINGTHATISMOSTLYOPAQUE;FOOBLA',
+    'date': 'Wed Jul 18 11:50:45 2012'},
+  {':method': "get",
+    ':path': '/generate_foo.html',
+    ':version': 'HTTP/1.1',
+    ':host': 'www.foo.com',
+    'date': 'Wed Jul 18 11:50:45 2012'},
+  {':method': "get",
+    ':path': '/index.css',
+    ':version': 'HTTP/1.1',
+    'user-agent': 'blah blah browser version blah blah',
+    'accept-encoding': 'sdch, bzip, compress',
+    ':host': 'www.foo.com',
+    'cookie': 'SOMELONGSTRINGTHATISMOSTLYOPAQUE;BLAJBLA',
+    'date': 'Wed Jul 18 11:50:46 2012'},
+  ]
+default_responses = [{':status': '200',
+                      ':status-text': 'ok',
+                      ':version': '1.1',
+                      ':scheme': 'https'} for x in default_requests]
+
